# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_ProTrack-Solutions_protrack-api_internal_domain_enums.Status:
    enum:
    - ACTIVE
    - INACTIVE
    - DELETED
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusDeleted
...
  github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.CreateProductCategoryRequest:
    properties:
      color:
        type: string
      name:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.ProductCategoryResponse:
    properties:
      color:
        type: string
      company_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_domain_enums.Status'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.SetProductCategoryStatusRequest:
    properties:
      id:
        type: string
      status:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_domain_enums.Status'
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.UpdateProductCategoryRequest:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
      updated_by:
        type: string
    type: object
...
  github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.CreateVendorsRequest:
    properties:
      address_line_1:
        type: string
      address_line_2:
        type: string
      city:
        type: string
      company_id:
        type: string
      country:
        type: string
      email:
        type: string
      name:
        type: string
      neighborhood:
        type: string
      number:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      state:
        type: string
      tax_id:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.ToggleVendorsActiveParams:
    properties:
      is_active:
        type: boolean
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.UpdateVendorsRequest:
    properties:
      address_line_1:
        type: string
      address_line_2:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      name:
        type: string
      neighborhood:
        type: string
      number:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      state:
        type: string
      tax_id:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.VendorResponse:
    properties:
      address_line_1:
        type: string
      address_line_2:
        type: string
      city:
        type: string
      company_id:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      neighborhood:
        type: string
      number:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      state:
        type: string
      tax_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: suporte@protrack.com
    name: Suporte ProTrack
  description: API de gerenciamento financeiro e comercial ProTrack Solutions.
  termsOfService: http://swagger.io/terms/
  title: ProTrack API
  version: "1.0"
paths:
...
  /products-categories:
    post:
      consumes:
      - application/json
      parameters:
      - description: Categoria
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.CreateProductCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria uma categoria de produto
      tags:
      - products-categories
  /products-categories/{id}:
    delete:
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Remove uma categoria de produto
      tags:
      - products-categories
    get:
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.ProductCategoryResponse'
      security:
      - BearerAuth: []
      summary: Busca categoria de produto por ID
      tags:
      - products-categories
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Categoria
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.UpdateProductCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.ProductCategoryResponse'
      security:
      - BearerAuth: []
      summary: Atualiza uma categoria de produto
      tags:
      - products-categories
  /products-categories/list/company:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.ProductCategoryResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista categorias de produto da empresa
      tags:
      - products-categories
  /products-categories/status/:
    put:
      consumes:
      - application/json
      parameters:
      - description: Status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_products_categories_domain.SetProductCategoryStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Altera o status de uma categoria de produto
      tags:
      - products-categories
...
  /vendors:
    post:
      consumes:
      - application/json
      parameters:
      - description: Fornecedor
        in: body
        name: vendor
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.CreateVendorsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Cria um fornecedor
      tags:
      - vendors
  /vendors/{id}:
    get:
      parameters:
      - description: ID do fornecedor
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.VendorResponse'
      security:
      - BearerAuth: []
      summary: Busca fornecedor por ID
      tags:
      - vendors
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do fornecedor
        in: path
        name: id
        required: true
        type: string
      - description: Fornecedor
        in: body
        name: vendor
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.UpdateVendorsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Atualiza um fornecedor
      tags:
      - vendors
  /vendors/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.VendorResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista fornecedores da empresa
      tags:
      - vendors
  /vendors/list/is-active:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.VendorResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista fornecedores ativos
      tags:
      - vendors
  /vendors/toggle/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do fornecedor
        in: path
        name: id
        required: true
        type: string
      - description: Status
        in: body
        name: toggle
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_vendors_domain.ToggleVendorsActiveParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Ativa ou desativa um fornecedor
      tags:
      - vendors
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"